Moved swr files from common/ to data/
