+++ react-ui/src/data/useSwrHallOfFame.jsx
@@ -2,7 +2,7 @@
  * Fetch winning flags from each contest
  */
 
-import { useSwrAuth } from '../common';
+import useSwrAuth from './useSwrAuth';
 
 const useSwrHallOfFame = () => useSwrAuth('/hallOfFame');
