+++ react-ui/src/data/useSwrModReview.jsx
@@ -2,7 +2,7 @@
  * Fetch winning flags from each contest
  */
 
-import { useSwrAuth } from '../common';
+import useSwrAuth from './useSwrAuth';
 
 const useSwrModReview = () => useSwrAuth('/mod/reviewSubmissions');
