+++ react-ui/src/data/useSwrSubmission.jsx
@@ -2,7 +2,7 @@
  * Fetch winning flags from each contest
  */
 
-import { useSwrAuth } from '../common';
+import useSwrAuth from './useSwrAuth';
 
 const useSwrSubmission = () => useSwrAuth('/submission');
