+++ react-ui/src/pages/settings/Settings.jsx
@@ -6,12 +6,13 @@ import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
 import { useLocation } from 'react-router-dom';
 
-import { useSnackbarState, useSwrMutation } from '../../common';
+import { useSnackbarState } from '../../common';
 import snackbarTypes from '../../common/snackbarTypes';
 import {
   CustomSwitch, Header, PageContainer, ProtectedRoute,
 } from '../../components';
 import { putData } from '../../data/api';
+import useSwrMutation from '../../data/useSwrMutation';
 
 const URL = '/settings';
