+++ react-ui/src/pages/submission/SubmissionForm.jsx
@@ -19,7 +19,6 @@ import {
   useAuthState,
   useFormState,
   useSnackbarState,
-  useSwrMutation,
 } from '../../common';
 import snackbarTypes from '../../common/snackbarTypes';
 import {
